Add new CHERI-Toooba bitstream
Blocks ifetch to non-memory-like addresses, so speculatively-executed
branches don't jump to bogus addresses with side-effects or that wedge
the interconnect (also applies to non-speculative branches, but normal
code shouldn't do that). In particular, an incorrectly-speculated branch
in some of the PPAC tests was redirecting the fetch unit to the PCI
bridge block's memory, but that IP appears to be of poor quality and
never delivered an AXI response, wedging the core.